Huge rework of the data structure
The data is now split into Tasks and Days, where days use TaskIds
to reference objects in the Tasks map. All of this is stored in
Database.

In addition the time logging logic is reworked, so now each day
has a Log which tracks all times, instead of each Focus tracking
its own breaks. Breaks and Focuses now happen at the same level.
Some commands are changed as a result of this.

Because Day doesn't have the Task names it can't handle Display.
We now have a DisplayDay struct for this.

edit_day_guard.rs and data.rs are no longer needed with the new
approach.

Year is removed but replaced with storage::YearData which is the
intermediary between the data model and the storage model.